% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethPed_checkNA.R
\name{checkNA}
\alias{checkNA}
\title{Check missing valus in dataset}
\usage{
checkNA(data)
}
\arguments{
\item{data}{Object in class data.frame or matrix or ExpressionSet.}
}
\description{
Check missing valus in dataset
}
\examples{
#################### Loading and view sample data
data(MethPed_sample)
head(MethPed_sample,10)

################### Check missing value
checkNA(MethPed_sample)
}

