% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethPed_plot.methped.R
\name{plot.methped}
\alias{plot.methped}
\title{Plot conditional probability of samples that belongs to different tumor subtypes.}
\usage{
\method{plot}{methped}(x, ...)
}
\arguments{
\item{x}{Object in "methped" class. Output of function \link[MethPed]{MethPed}.}

\item{...}{More arguments from function \link[graphics]{barplot}.}
}
\value{
Object in "methped" class. Output of function \link[MethPed]{MethPed}.
}
\description{
Plot conditional probability of samples that belongs to different tumor subtypes.
}
\examples{

#################### Loading sample data
data(MethPed_sample)

#################### Applying MethPed to sample data
res<-MethPed(MethPed_sample)

#################### Plot conditional probability
plot(res)

}

