% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethPed_probeMis.R
\name{probeMis}
\alias{probeMis}
\title{Missing probe names from training data compared with input samples.}
\usage{
probeMis(x)
}
\arguments{
\item{x}{Object in MethPed class. Output of function \link[MethPed]{MethPed}.}
}
\value{
Object in "methped" class. Output of function \link[MethPed]{MethPed}.
}
\description{
Missing probe names from training data compared with input samples.
}
\examples{

#################### Loading sample data
data(MethPed_sample)

#################### Applying MethPed to sample data
res<-MethPed(MethPed_sample)

#################### The names of the probes that were included in the original
                   # classifier but are missing from the data at hand
probeMis(res)

}

