% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MethPed_summary.methped.R
\name{summary.methped}
\alias{summary.methped}
\title{Summary of conditional probability or binary classification of samples that belong to different tumor subtypes.}
\usage{
\method{summary}{methped}(object, ...)
}
\arguments{
\item{object}{Object in methped class. Output of function \link[MethPed]{MethPed}.}

\item{...}{Additional arguments affecting the summary produced}
}
\value{
Object in "methped" class. Output of function \link[MethPed]{MethPed}.
}
\description{
Summary of conditional probability or binary classification of samples that belong to different tumor subtypes.
}
\examples{

#################### Loading sample data
data(MethPed_sample)

#################### Applying MethPed to sample data
res<-MethPed(MethPed_sample)

#################### Summary function of MethPed output
summary (res)

}

